+++ ogma-core/src/Command/ROSApp.hs
@@ -60,6 +60,7 @@ import Command.VariableDB (Connection (..), InputDef (..), TopicDef (..),
                            findTopic, findType, findTypeByType)
 import Data.Aeson.Extra   (mergeObjects)
 import Data.ExprPair      (ExprPair(..), exprPair)
+import Data.Location      (Location (..))
 import Data.Spec.Parser   (readInputExpr)
 
 -- | Generate a new ROS application connected to Copilot.
@@ -96,13 +97,20 @@ command' options (ExprPair exprT) = do
     varDB <- openVarDBFilesWithDefault varDBFile
 
     specT <- maybe (return Nothing) (\e -> Just . InputFileSpec <$> readInputExpr' e) cExpr
-    specF <- maybe (return Nothing) (\f -> Just <$> readInputFile' f) fp
+    specF <- if null fpA
+                  then return Nothing
+                  else do
+                    fpA' <- mapM readInputFile' fpA
+                    let fpA'' = combineInputFiles fpA'
+                    if length fpA'' > 1
+                      then liftEither $ Left commandMultipleInputTypes
+                      else pure $ Just $ head fpA''
 
     let spec = specT <|> specF
 
     liftEither $ checkArguments spec vs rs
 
-    copilotM <- sequenceA $ (\spec' -> processSpec spec' cExpr fp) <$> spec
+    copilotM <- sequenceA $ (\spec' -> processSpec spec' cExpr fpA) <$> spec
 
     let varNames = fromMaybe (defaultVarNames spec) vs
         monitors = maybe (defaultMonitors spec) (map (\x -> (x, Nothing))) rs
@@ -124,7 +132,7 @@ command' options (ExprPair exprT) = do
   where
 
     cExpr          = commandConditionExpr options
-    fp             = commandInputFile options
+    fpA            = commandInputFiles options
     varNameFile    = commandVariables options
     varDBFile      = maybeToList $ commandVariableDB options
     handlersFile   = commandHandlers options
@@ -161,7 +169,7 @@ command' options (ExprPair exprT) = do
 -- applications.
 data CommandOptions = CommandOptions
   { commandConditionExpr :: Maybe String   -- ^ Trigger condition.
-  , commandInputFile   :: Maybe FilePath -- ^ Input specification file.
+  , commandInputFiles  :: [FilePath]     -- ^ Input specification files.
   , commandTargetDir   :: FilePath       -- ^ Target directory where the
                                          -- application should be created.
   , commandTemplateDir :: Maybe FilePath -- ^ Directory where the template is
@@ -280,3 +288,16 @@ randomBaseType ty = case ty of
   "float"    -> "randomFloat"
   "double"   -> "randomFloat"
   def        -> def
+
+-- | Error message associated to having multiple input files of incompatible
+-- types.
+commandMultipleInputTypes :: ErrorTriplet
+commandMultipleInputTypes =
+    ErrorTriplet ecMultipleInputTypes msg LocationNothing
+  where
+    msg =
+      "Too many inputs provided. Provide one diagram or multiple specs."
+
+-- | Error: multiple inputs of incompatible types.
+ecMultipleInputTypes :: ErrorCode
+ecMultipleInputTypes = 1